requirements.txt
tensorflow==2.17.1
keras>=3.2.0  
numpy==1.26.0
protobuf>=3.20,<4 
streamlit==1.25.0
Pillow
requests
